refactor: full popup animation
