+++ src/components/popup/Portal.tsx
@@ -1,13 +1,15 @@
 import { PropsWithChildren } from 'react';
 import { createPortal } from 'react-dom';
 
+import { AnimatePresence } from 'framer-motion';
 import { useIsMounted } from 'usehooks-ts';
 
 type PortalProps = {
   elementId?: string;
+  show: boolean;
 };
 
-export default function Portal({ elementId = 'popup-portal', children }: PropsWithChildren<PortalProps>) {
+export default function Portal({ elementId = 'normal-portal-root', show, children }: PropsWithChildren<PortalProps>) {
   const isMounted = useIsMounted();
 
   if (!isMounted) {
@@ -20,5 +22,5 @@ export default function Portal({ elementId = 'popup-portal', children }: PropsWi
     return null;
   }
 
-  return createPortal(children, portalRoot);
+  return createPortal(<AnimatePresence mode="wait">{show && children}</AnimatePresence>, portalRoot);
 }
